Mochi ai scout/requirements.txt
google-genai
feedparser
requests